git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@9651 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/segment.h
@@ -239,11 +239,11 @@ class CSide {
 		bool IsTextured (void);
 
 	private:
-		void SetupCorners (short* verts, int* index);
-		void SetupVertexList (short* verts, int* index);
+		void SetupCorners (short* verts, short* index);
+		void SetupVertexList (short* verts, short* index);
 		void SetupFaceVertIndex (void);
-		void SetupAsQuad (CFixVector& vNormal, CFloatVector& vNormalf, short* verts, int* index);
-		void SetupAsTriangles (bool bSolid, short* verts, int* index);
+		void SetupAsQuad (CFixVector& vNormal, CFloatVector& vNormalf, short* verts, short* index);
+		void SetupAsTriangles (bool bSolid, short* verts, short* index);
 	};
 
 //------------------------------------------------------------------------------
